move communication service to domain layer
