+++ p2p/infra/adapter/grpc_command_service.go
@@ -28,7 +28,6 @@ import (
 //kind of error
 var ErrEmptyPeerId = errors.New("empty nodeid proposed")
 var ErrEmptyLeaderId = errors.New("empty leader id proposed")
-var ErrEmptyConnectionId = errors.New("empty connection ")
 
 type Publish func(exchange string, topic string, data interface{}) (err error) // 나중에 의존성 주입을 해준다.
